+++ src/clientanalytics.js
@@ -1,46 +1,65 @@
-const CA_COLLECTION = "enlearnEventLog";
-const CA_QUERY_ALL = "getAllEvents";
-const CA_QUERY_LATEST = "getLatestEvent";
+const EVENTS_COLLECTION = "enlearnEventLog";
+const QUERY_ALL = "getAllEvents";
+const QUERY_NOT_UPLOADED = "getEventsToUpload";
+const QUERY_MARK_UPLOADED = "markEventsAsUploaded";
 
-export class ClientAnalyticsEventLogStore {
-  constructor(clientAnalytics) {
-    this._ca = clientAnalytics;
-  }
+export const createClientAnalyticsEventLogStore = clientAnalytics => {
+  const getAllEvents = () =>
+    clientAnalytics.query(QUERY_ALL, {}, EVENTS_COLLECTION);
+  const recordEvent = event =>
+    clientAnalytics.insert(EVENTS_COLLECTION, { event, uploaded: false });
+  const getEventsToUpload = () =>
+    clientAnalytics.query(QUERY_NOT_UPLOADED, {}, EVENTS_COLLECTION);
+  const markEventsAsUploaded = ids =>
+    clientAnalytics.query(QUERY_MARK_UPLOADED, { ids }, EVENTS_COLLECTION);
 
-  initialize() {
-    return this._ca
-      .createCollection(CA_COLLECTION)
-      .then(() =>
-        this._ca.registerQuery(CA_QUERY_ALL, collection => {
-          return collection
+  const store = {
+    getAllEvents,
+    recordEvent,
+    getEventsToUpload,
+    markEventsAsUploaded,
+  };
+
+  return clientAnalytics
+    .createCollection(EVENTS_COLLECTION)
+    .then(() =>
+      clientAnalytics.registerQuery(QUERY_ALL, collection => {
+        return collection
+          .chain()
+          .simplesort("event.sequenceNumber", false)
+          .data()
+          .map(r => r.event);
+      })
+    )
+    .then(() =>
+      clientAnalytics.registerQuery(QUERY_NOT_UPLOADED, collection => {
+        return collection
+          .chain()
+          .find({ uploaded: { $ne: true } })
+          .limit(50)
+          .data()
+          .map(r => r.event);
+      })
+    )
+    .then(() =>
+      clientAnalytics.registerQuery(
+        QUERY_MARK_UPLOADED,
+        (collection, params) => {
+          const records = collection
             .chain()
-            .simplesort("event.sequenceNumber", false)
-            .data()
-            .map(r => r.event);
-        })
+            .find({ uploaded: { $ne: true } })
+            .where(r => params.ids.indexOf(r.event.id) >= 0)
+            .data();
+
+          for (let i = 0; i < records.length; i++) {
+            const r = records[i];
+            r.uploaded = true;
+            collection.update(r);
+          }
+
+          return null;
+        }
       )
-      .then(() =>
-        this._ca.registerQuery(CA_QUERY_LATEST, collection => {
-          const results = collection
-            .chain()
-            .simplesort("event.sequenceNumber", true)
-            .limit(1)
-            .data()
-            .map(r => r.event);
-          return results.length > 0 ? results[0] : null;
-        })
-      );
-  }
-
-  getAllEvents() {
-    return this._ca.query(CA_QUERY_ALL, {}, CA_COLLECTION);
-  }
-
-  getLatestEvent() {
-    return this._ca.query(CA_QUERY_LATEST, {}, CA_COLLECTION);
-  }
-
-  recordEvent(event) {
-    return this._ca.insert(CA_COLLECTION, { event });
-  }
-}
+    )
+    .then(() => store);
+};