+++ src/userdata.test.js
@@ -1,6 +1,6 @@
-import { UserDataEventLogStore } from "./userdata";
+import { createUserDataEventLogStore } from "./userdata";
 
-function setup() {
+const setup = async () => {
   const userData = {
     read: jest.fn().mockImplementation((key, cb) => {
       cb();
@@ -10,71 +10,79 @@ function setup() {
     }),
   };
 
-  const testObj = new UserDataEventLogStore(userData);
-
+  const testObj = await createUserDataEventLogStore(userData);
   return { testObj, userData };
-}
+};
 
-test("records starts empty", () => {
-  const { testObj } = setup();
-  return expect(testObj.getLatestEvent()).resolves.toEqual(null);
+test("initialize reads from user data", async () => {
+  const callbacks = {};
+
+  const userData = {
+    read: jest.fn().mockImplementation((key, callback) => {
+      callbacks[key] = callback;
+    }),
+    write: jest.fn(),
+  };
+
+  const promise = createUserDataEventLogStore(userData);
+
+  const events = [{ id: 1 }, { id: 2 }, { id: 3 }];
+  callbacks["enlearnEventLog"](events);
+  callbacks["enlearnEventLogNotUploaded"]([]);
+
+  const testObj = await promise;
+  expect(testObj.getAllEvents()).resolves.toEqual(events);
 });
 
-test("initialize reads from user data", () => {
-  const { testObj, userData } = setup();
-
-  userData.read.mockReset();
-
-  const promise = testObj.initialize();
-
-  expect(userData.read).toHaveBeenCalledWith(
-    "enlearnEventLog",
-    expect.any(Function)
-  );
-
-  const callback = userData.read.mock.calls[0][1];
-
-  const events = [
-    { type: "a", value: 1 },
-    { type: "a", value: 2 },
-    { type: "a", value: 3 },
-  ];
-  callback(events);
-
-  return promise.then(() => {
-    return expect(testObj.getAllEvents()).resolves.toEqual(events);
-  });
+test("records starts empty", async () => {
+  const { testObj } = await setup();
+  return expect(testObj.getAllEvents()).resolves.toEqual([]);
 });
 
-test("recordEvent adds a record", () => {
-  const { testObj } = setup();
+test("recordEvent adds a record", async () => {
+  const { testObj } = await setup();
 
-  const event = { type: "a", value: "blah" };
+  const event = { id: 1, type: "a", value: "blah" };
   return testObj
     .recordEvent(event)
-    .then(() => expect(testObj.getLatestEvent()).resolves.toEqual(event));
+    .then(() => expect(testObj.getAllEvents()).resolves.toEqual([event]));
 });
 
 test("recordEvent writes records to store", async () => {
-  const { testObj, userData } = setup();
+  const { testObj, userData } = await setup();
 
-  await testObj.recordEvent(1);
+  await testObj.recordEvent({ id: 1 });
   expect(userData.write).toHaveBeenCalledWith(
     "enlearnEventLog",
+    [{ id: 1 }],
+    expect.any(Function)
+  );
+  expect(userData.write).toHaveBeenCalledWith(
+    "enlearnEventLogNotUploaded",
     [1],
     expect.any(Function)
   );
 
-  await testObj.recordEvent(2);
+  await testObj.recordEvent({ id: 2 });
   expect(userData.write).toHaveBeenCalledWith(
     "enlearnEventLog",
+    [{ id: 1 }, { id: 2 }],
+    expect.any(Function)
+  );
+  expect(userData.write).toHaveBeenCalledWith(
+    "enlearnEventLogNotUploaded",
     [1, 2],
     expect.any(Function)
   );
 
-  await testObj.recordEvent(3);
+  await testObj.recordEvent({ id: 3 });
   expect(userData.write).toHaveBeenCalledWith(
     "enlearnEventLog",
+    [{ id: 1 }, { id: 2 }, { id: 3 }],
+    expect.any(Function)
+  );
+  expect(userData.write).toHaveBeenCalledWith(
+    "enlearnEventLogNotUploaded",
     [1, 2, 3],
     expect.any(Function)
   );